backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
python-dotenv
openai
requests